news-producer/requirements.txt
aniso8601==8.1.0
asgiref==3.4.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
envyaml==1.9.210927
fastapi==0.68.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.1
h11==0.12.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.1
kafka-python==2.0.2
MarkupSafe==1.1.1
marshmallow==3.10.0
mongo==0.2.0
nose==1.3.7
numpy==1.20.1
pandas==1.2.2
pydantic==1.8.2
pymongo==3.12.0
python-dateutil==2.8.1
python-dotenv==0.19.0
pytz==2020.5
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
schedule==1.1.0
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.22
starlette==0.14.2
threadpoolctl==2.1.0
typing-extensions==3.10.0.2
urllib3==1.26.3
uvicorn==0.15.0
Werkzeug==1.0.1
